// src/lib/api/generated/model/invitationStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type InvitationStatus = typeof InvitationStatus[keyof typeof InvitationStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvitationStatus = {
  pending: 'pending',
  accepted: 'accepted',
  expired: 'expired',
  cancelled: 'cancelled',
} as const;